Authentication and QueryConfig Controllers

+++ DotNet/DataAcquisition/Application/Commands/Config/QueryConfig/DeleteFhirQueryConfigurationCommand.cs
@@ -1,5 +1,5 @@
 ﻿using LantanaGroup.Link.DataAcquisition.Application.Interfaces;
-using LantanaGroup.Link.DataAcquisition.Application.Repositories;
+using LantanaGroup.Link.DataAcquisition.Application.Models.Exceptions;
 using MediatR;
 
 namespace LantanaGroup.Link.DataAcquisition.Application.Commands.Config.QueryConfig;
@@ -24,7 +24,13 @@ public class DeleteFhirQueryConfigurationCommandHandler : IRequestHandler<Delete
 
     public async Task<Unit> Handle(DeleteFhirQueryConfigurationCommand request, CancellationToken cancellationToken)
     {
+        if (string.IsNullOrWhiteSpace(request.FacilityId))
+        {
+            throw new BadRequestException("GetFhirQueryConfigQuery.FacilityId is null or empty.");
+        }
+
         await _repository.DeleteAsync(request.FacilityId, cancellationToken);
+
         return new Unit();
     }
 }
